@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -17,23 +17,23 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.3
   provider: ^5.0.0
-  random_words: ^1.0.2
+  english_words: ^4.0.0
   # intl: ^0.16.1
-  intl: ^0.17.0-nullsafety.2
-  bloc: ^7.0.0
-  flutter_bloc: ^7.0.1
-  shared_preferences: ^2.0.6
-  data_tables: ^1.0.3
+  intl: ^0.17.0
+  bloc: ^7.2.1
+  flutter_bloc: ^7.1.0
+  shared_preferences: ^2.0.8
+  data_tables: ^1.4.0
   timeline_tile: ^2.0.0
   # device_preview: ^0.5.5
   device_preview: ^0.7.4
-  signalr_core: ^1.0.5
+  signalr_core: ^1.1.1
   # audioplayers: ^0.19.1
-  share_plus: ^2.1.4
+  share_plus: ^3.0.2
   share_plus_web: ^2.0.4
-  url_launcher: ^6.0.9
+  url_launcher: ^6.0.12
   social_share: ^2.1.1
-  get_it: ^7.1.3
+  get_it: ^7.2.0
 
 
 dev_dependencies:
