@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:test_app/components/card/level2.dart';
 
 class Level1 extends StatefulWidget {
-  final String title;
+  final String? title;
   Level1({this.title});
 
   @override
