@@ -1,7 +1,7 @@
 class User {
-  final String title;
-  final String role;
-  final DateTime joiningDate;
+  final String? title;
+  final String? role;
+  final DateTime? joiningDate;
 
   User({this.title, this.role, this.joiningDate});
 }
