@@ -1,12 +1,12 @@
 import 'package:meta/meta.dart';
 
 class Weather {
-  final String cityName;
+  final String? cityName;
   final double temperatureCelsius;
 
   Weather({
-    @required this.cityName,
-    @required this.temperatureCelsius,
+    required this.cityName,
+    required this.temperatureCelsius,
   });
 
   @override
