@@ -5,9 +5,9 @@ import 'package:test_app/models/data.dart';
 class DataProvider extends ChangeNotifier {
   Data _data = new Data(title: 'Base title', color: null);
 
-  String get getTitle => _data.title;
+  String? get getTitle => _data.title;
 
-  Color get getColor => _data.color;
+  Color? get getColor => _data.color;
 
   void changeTitle(String newTitle) {
     _data.setData(newTitle);
