@@ -4,12 +4,12 @@ import 'package:test_app/models/weather.dart';
 
 abstract class WeatherRepository {
   /// Throws [NetworkException].
-  Future<Weather> fetchWeather(String cityName);
+  Future<Weather> fetchWeather(String? cityName);
 }
 
 class FakeWeatherRepository implements WeatherRepository {
   @override
-  Future<Weather> fetchWeather(String cityName) {
+  Future<Weather> fetchWeather(String? cityName) {
     // Simulate network delay
     return Future.delayed(
       Duration(seconds: 1),
